refactor: スレッド書き込みテーブルのファクトリを1つにまとめる
